Refactoring test package structure * Moves test into mirrored hierarchy with main * Separates QuestionSet tests from Person tests
