+++ src/test/java/com/blockscore/SearchTest.java
