+++ src/test/java/com/blockscore/WatchlistTest.java
